feat: added usage stats support
